+++ tests/genai/judges/adapters/test_databricks_adapter.py
@@ -11,13 +11,15 @@ from mlflow.entities.trace_info import TraceInfo
 from mlflow.entities.trace_location import TraceLocation
 from mlflow.entities.trace_state import TraceState
 from mlflow.exceptions import MlflowException
-from mlflow.genai.judges.adapters.databricks_adapter import (
+from mlflow.genai.judges.adapters.databricks_managed_judge_adapter import (
+    call_chat_completions,
+)
+from mlflow.genai.judges.adapters.databricks_serving_endpoint_adapter import (
     InvokeDatabricksModelOutput,
     _invoke_databricks_serving_endpoint,
     _parse_databricks_model_response,
     _record_judge_model_usage_failure_databricks_telemetry,
     _record_judge_model_usage_success_databricks_telemetry,
-    call_chat_completions,
 )
 from mlflow.utils import AttrDict
 
@@ -123,7 +125,7 @@ def test_invoke_databricks_serving_endpoint_successful_invocation() -> None:
             return_value=mock_creds,
         ) as mock_get_creds,
         mock.patch(
-            "mlflow.genai.judges.adapters.databricks_adapter.requests.post",
+            "mlflow.genai.judges.adapters.databricks_serving_endpoint_adapter.requests.post",
             return_value=mock_response,
         ) as mock_post,
     ):
@@ -160,7 +162,7 @@ def test_invoke_databricks_serving_endpoint_bad_request_error_no_retry(status_co
             return_value=mock_creds,
         ) as mock_get_creds,
         mock.patch(
-            "mlflow.genai.judges.adapters.databricks_adapter.requests.post",
+            "mlflow.genai.judges.adapters.databricks_serving_endpoint_adapter.requests.post",
             return_value=mock_response,
         ) as mock_post,
     ):
@@ -194,10 +196,12 @@ def test_invoke_databricks_serving_endpoint_retry_logic_with_transient_errors() 
             return_value=mock_creds,
         ) as mock_get_creds,
         mock.patch(
-            "mlflow.genai.judges.adapters.databricks_adapter.requests.post",
+            "mlflow.genai.judges.adapters.databricks_serving_endpoint_adapter.requests.post",
             side_effect=[error_response, success_response],
         ) as mock_post,
-        mock.patch("mlflow.genai.judges.adapters.databricks_adapter.time.sleep") as mock_sleep,
+        mock.patch(
+            "mlflow.genai.judges.adapters.databricks_serving_endpoint_adapter.time.sleep"
+        ) as mock_sleep,
     ):
         result = _invoke_databricks_serving_endpoint(
             model_name="test-model", prompt="test prompt", num_retries=3
@@ -225,7 +229,7 @@ def test_invoke_databricks_serving_endpoint_json_decode_error() -> None:
             return_value=mock_creds,
         ) as mock_get_creds,
         mock.patch(
-            "mlflow.genai.judges.adapters.databricks_adapter.requests.post",
+            "mlflow.genai.judges.adapters.databricks_serving_endpoint_adapter.requests.post",
             return_value=mock_response,
         ) as mock_post,
     ):
@@ -247,10 +251,12 @@ def test_invoke_databricks_serving_endpoint_connection_error_with_retries() -> N
             return_value=mock_creds,
         ) as mock_get_creds,
         mock.patch(
-            "mlflow.genai.judges.adapters.databricks_adapter.requests.post",
+            "mlflow.genai.judges.adapters.databricks_serving_endpoint_adapter.requests.post",
             side_effect=requests.ConnectionError("Connection failed"),
         ) as mock_post,
-        mock.patch("mlflow.genai.judges.adapters.databricks_adapter.time.sleep") as mock_sleep,
+        mock.patch(
+            "mlflow.genai.judges.adapters.databricks_serving_endpoint_adapter.time.sleep"
+        ) as mock_sleep,
     ):
         with pytest.raises(
             MlflowException, match="Failed to invoke Databricks model after 3 attempts"
@@ -292,7 +298,8 @@ def test_invoke_databricks_serving_endpoint_with_response_schema():
 
     with (
         mock.patch(
-            "mlflow.genai.judges.adapters.databricks_adapter.requests.post", side_effect=mock_post
+            "mlflow.genai.judges.adapters.databricks_serving_endpoint_adapter.requests.post",
+            side_effect=mock_post,
         ),
         mock.patch(
             "mlflow.utils.databricks_utils.get_databricks_host_creds", return_value=mock_creds
@@ -460,10 +467,12 @@ def mock_databricks_rag_eval():
 def test_call_chat_completions_success(user_prompt, system_prompt, mock_databricks_rag_eval):
     with (
         mock.patch(
-            "mlflow.genai.judges.adapters.databricks_adapter._check_databricks_agents_installed"
+            "mlflow.genai.judges.adapters.databricks_managed_judge_adapter._check_databricks_agents_installed"
         ),
         mock.patch.dict("sys.modules", {"databricks.rag_eval": mock_databricks_rag_eval["module"]}),
-        mock.patch("mlflow.genai.judges.adapters.databricks_adapter.VERSION", "1.0.0"),
+        mock.patch(
+            "mlflow.genai.judges.adapters.databricks_managed_judge_adapter.VERSION", "1.0.0"
+        ),
     ):
         result = call_chat_completions(user_prompt, system_prompt)
 
@@ -485,10 +494,12 @@ def test_call_chat_completions_with_custom_session_name(mock_databricks_rag_eval
     """Test that custom session name is used when provided."""
     with (
         mock.patch(
-            "mlflow.genai.judges.adapters.databricks_adapter._check_databricks_agents_installed"
+            "mlflow.genai.judges.adapters.databricks_managed_judge_adapter._check_databricks_agents_installed"
         ),
         mock.patch.dict("sys.modules", {"databricks.rag_eval": mock_databricks_rag_eval["module"]}),
-        mock.patch("mlflow.genai.judges.adapters.databricks_adapter.VERSION", "1.0.0"),
+        mock.patch(
+            "mlflow.genai.judges.adapters.databricks_managed_judge_adapter.VERSION", "1.0.0"
+        ),
     ):
         custom_session_name = "custom-session-name"
         result = call_chat_completions(
@@ -516,7 +527,7 @@ def test_call_chat_completions_client_error(mock_databricks_rag_eval):
 
     with (
         mock.patch(
-            "mlflow.genai.judges.adapters.databricks_adapter._check_databricks_agents_installed"
+            "mlflow.genai.judges.adapters.databricks_managed_judge_adapter._check_databricks_agents_installed"
         ),
         mock.patch.dict("sys.modules", {"databricks.rag_eval": mock_databricks_rag_eval["module"]}),
     ):
